+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/jss/compare/v1.1.2...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- Returns ``404: Not Found`` for non-existing user
 
 [1.1.2](https://github.com/jshwi/jss/releases/tag/v1.1.2) - 2021-10-05
 ------------------------------------------------------------------------